Refactor into new files
